+++ Packages/com.chark.game-management/Runtime/Messaging/Delegates.cs
@@ -1,0 +1,4 @@
+﻿namespace CHARK.GameManagement.Messaging
+{
+    public delegate void OnMessageReceived<in TMessage>(TMessage message) where TMessage : IMessage;
+}