+++ Packages/com.chark.game-management/CHANGELOG.md
@@ -25,6 +25,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - `GameStorage.GetValueAsync` (now `GameStorage.ReadValueAsync`) not switching back to main thread when no value is found.
+- Message bus not handling abstract and interface listeners.
+- Message bus iteration breaking when listener would remove itself during `Raise` call.
 
 ## [v0.0.2](https://github.com/chark/game-management/compare/v0.0.1...v0.0.2) - 2023-10-06